refactor(database): isolate catalog adapter
