+++ crates/coral-engine/src/backends/database/mod.rs
@@ -1,7 +1,8 @@
 //! Relational database backend registration through `datafusion-table-providers`.
 
+mod catalog;
+
 use std::collections::{BTreeMap, BTreeSet, HashMap};
-use std::fmt;
 use std::future::Future;
 use std::sync::Arc;
 use std::time::Duration;
@@ -12,25 +13,19 @@ use coral_spec::backends::database::{
     SqliteConnectionSpec,
 };
 use coral_spec::{ParsedTemplate, SourceManifestCommon};
-use datafusion::arrow::array::StringArray;
-use datafusion::arrow::datatypes::{DataType, Field, Schema};
-use datafusion::catalog::{CatalogProvider, SchemaProvider};
-use datafusion::datasource::TableProvider;
 use datafusion::error::{DataFusionError, Result as DataFusionResult};
-use datafusion::logical_expr::TableType;
 use datafusion::prelude::SessionContext;
-use datafusion::sql::TableReference;
-use datafusion::sql::unparser::dialect::{Dialect, MySqlDialect, PostgreSqlDialect, SqliteDialect};
+use datafusion::sql::unparser::dialect::{MySqlDialect, PostgreSqlDialect, SqliteDialect};
 use datafusion_table_providers::UnsupportedTypeAction;
-use datafusion_table_providers::sql::db_connection_pool::dbconnection::query_arrow;
+use datafusion_table_providers::sql::db_connection_pool::Mode;
 use datafusion_table_providers::sql::db_connection_pool::mysqlpool::MySQLConnectionPool;
 use datafusion_table_providers::sql::db_connection_pool::postgrespool::PostgresConnectionPool;
 use datafusion_table_providers::sql::db_connection_pool::sqlitepool::SqliteConnectionPoolFactory;
-use datafusion_table_providers::sql::db_connection_pool::{DbConnectionPool, Mode};
-use datafusion_table_providers::sql::sql_provider_datafusion::SqlTable;
 use datafusion_table_providers::util::secrets::to_secret_map;
-use futures::TryStreamExt as _;
 
+use self::catalog::{
+    DatabaseCatalog, DatabaseRelation, boxed_provider_error, build_database_catalog, provider_error,
+};
 use crate::backends::shared::template::{RenderContext, render_template};
 use crate::backends::{
     BackendCatalogRegistration, BackendCompileRequest, BackendRegistration,
@@ -184,7 +179,7 @@ impl DatabaseCatalogStrategy for PostgresConnectionSpec {
             .map_err(provider_error)?
             // The MySQL adapter has no equivalent unsupported-type policy.
             .with_unsupported_type_action(UnsupportedTypeAction::String);
-        database_catalog(
+        build_database_catalog(
             Arc::new(pool),
             POSTGRES_RELATIONS_SQL,
             Arc::new(PostgreSqlDialect {}),
@@ -226,7 +221,7 @@ impl DatabaseCatalogStrategy for MySqlConnectionSpec {
         let pool = MySQLConnectionPool::new(to_secret_map(params))
             .await
             .map_err(provider_error)?;
-        database_catalog(
+        build_database_catalog(
             Arc::new(pool),
             MYSQL_RELATIONS_SQL,
             Arc::new(MySqlDialect {}),
@@ -264,7 +259,7 @@ impl DatabaseCatalogStrategy for SqliteConnectionSpec {
             .build()
             .await
             .map_err(boxed_provider_error)?;
-        database_catalog(
+        build_database_catalog(
             Arc::new(pool),
             SQLITE_RELATIONS_SQL,
             Arc::new(SqliteDialect {}),
@@ -293,195 +288,6 @@ SELECT 'main' AS schema_name,
        CASE type WHEN 'view' THEN 'VIEW' ELSE 'BASE TABLE' END AS relation_type
 FROM sqlite_master
 WHERE type IN ('table', 'view') AND name NOT LIKE 'sqlite_%'";
-
-type Pool<T, P> = Arc<dyn DbConnectionPool<T, P> + Send + Sync>;
-
-struct DatabaseCatalog {
-    provider: Arc<dyn CatalogProvider>,
-    relations: Vec<DatabaseRelation>,
-}
-
-#[derive(Clone, Debug)]
-struct DatabaseRelation {
-    schema_name: String,
-    table_name: String,
-    table_type: TableType,
-}
-
-type SqlDialect = Arc<dyn Dialect + Send + Sync>;
-
-async fn database_catalog<T: 'static, P: 'static>(
-    pool: Pool<T, P>,
-    inventory_sql: &str,
-    dialect: SqlDialect,
-) -> DataFusionResult<DatabaseCatalog> {
-    let relations = load_database_inventory(&pool, inventory_sql).await?;
-    let provider = Arc::new(CoralDatabaseCatalogProvider::new(
-        &pool, &relations, &dialect,
-    )) as Arc<dyn CatalogProvider>;
-    Ok(DatabaseCatalog {
-        provider,
-        relations,
-    })
-}
-
-async fn load_database_inventory<T: 'static, P: 'static>(
-    pool: &Pool<T, P>,
-    inventory_sql: &str,
-) -> DataFusionResult<Vec<DatabaseRelation>> {
-    let schema = Arc::new(Schema::new(vec![
-        Field::new("schema_name", DataType::Utf8, false),
-        Field::new("table_name", DataType::Utf8, false),
-        Field::new("relation_type", DataType::Utf8, false),
-    ]));
-    let connection = pool.connect().await.map_err(boxed_provider_error)?;
-    let batches = query_arrow(connection, inventory_sql.to_string(), Some(schema))
-        .await
-        .map_err(provider_error)?
-        .try_collect::<Vec<_>>()
-        .await?;
-    let mut relations = Vec::new();
-    for batch in batches {
-        let schema_names = database_inventory_column(&batch, "schema_name")?;
-        let table_names = database_inventory_column(&batch, "table_name")?;
-        let relation_types = database_inventory_column(&batch, "relation_type")?;
-        for row in 0..batch.num_rows() {
-            relations.push(DatabaseRelation {
-                schema_name: schema_names.value(row).to_string(),
-                table_name: table_names.value(row).to_string(),
-                table_type: relation_table_type(relation_types.value(row)),
-            });
-        }
-    }
-    relations.sort_by(|left, right| {
-        (&left.schema_name, &left.table_name).cmp(&(&right.schema_name, &right.table_name))
-    });
-    relations.dedup_by(|left, right| {
-        left.schema_name == right.schema_name && left.table_name == right.table_name
-    });
-    Ok(relations)
-}
-
-fn database_inventory_column<'a>(
-    batch: &'a datafusion::arrow::record_batch::RecordBatch,
-    name: &str,
-) -> DataFusionResult<&'a StringArray> {
-    batch
-        .column_by_name(name)
-        .and_then(|column| column.as_any().downcast_ref::<StringArray>())
-        .ok_or_else(|| {
-            DataFusionError::Execution(format!("database inventory column '{name}' is not Utf8"))
-        })
-}
-
-fn relation_table_type(value: &str) -> TableType {
-    if value.eq_ignore_ascii_case("VIEW") || value.eq_ignore_ascii_case("MATERIALIZED VIEW") {
-        TableType::View
-    } else {
-        TableType::Base
-    }
-}
-
-struct CoralDatabaseCatalogProvider {
-    schemas: HashMap<String, Arc<dyn SchemaProvider>>,
-}
-
-impl CoralDatabaseCatalogProvider {
-    fn new<T: 'static, P: 'static>(
-        pool: &Pool<T, P>,
-        relations: &[DatabaseRelation],
-        dialect: &SqlDialect,
-    ) -> Self {
-        let mut by_schema = BTreeMap::<String, BTreeMap<String, TableType>>::new();
-        for relation in relations {
-            by_schema
-                .entry(relation.schema_name.clone())
-                .or_default()
-                .insert(relation.table_name.clone(), relation.table_type);
-        }
-        let schemas = by_schema
-            .into_iter()
-            .map(|(schema_name, tables)| {
-                let provider = CoralDatabaseSchemaProvider {
-                    schema_name: schema_name.clone(),
-                    tables,
-                    pool: Arc::clone(pool),
-                    dialect: Arc::clone(dialect),
-                };
-                (schema_name, Arc::new(provider) as Arc<dyn SchemaProvider>)
-            })
-            .collect();
-        Self { schemas }
-    }
-}
-
-impl fmt::Debug for CoralDatabaseCatalogProvider {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        formatter
-            .debug_struct("CoralDatabaseCatalogProvider")
-            .field("schemas", &self.schemas.keys().collect::<Vec<_>>())
-            .finish()
-    }
-}
-
-impl CatalogProvider for CoralDatabaseCatalogProvider {
-    fn schema_names(&self) -> Vec<String> {
-        self.schemas.keys().cloned().collect()
-    }
-
-    fn schema(&self, name: &str) -> Option<Arc<dyn SchemaProvider>> {
-        self.schemas.get(name).cloned()
-    }
-}
-
-struct CoralDatabaseSchemaProvider<T: 'static, P: 'static> {
-    schema_name: String,
-    tables: BTreeMap<String, TableType>,
-    pool: Pool<T, P>,
-    dialect: SqlDialect,
-}
-
-impl<T: 'static, P: 'static> fmt::Debug for CoralDatabaseSchemaProvider<T, P> {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        formatter
-            .debug_struct("CoralDatabaseSchemaProvider")
-            .field("schema_name", &self.schema_name)
-            .field("tables", &self.tables)
-            .finish_non_exhaustive()
-    }
-}
-
-#[async_trait]
-impl<T: 'static, P: 'static> SchemaProvider for CoralDatabaseSchemaProvider<T, P> {
-    fn table_names(&self) -> Vec<String> {
-        self.tables.keys().cloned().collect()
-    }
-
-    async fn table(&self, name: &str) -> DataFusionResult<Option<Arc<dyn TableProvider>>> {
-        if !self.tables.contains_key(name) {
-            return Ok(None);
-        }
-        SqlTable::new(
-            &self.schema_name,
-            &self.pool,
-            TableReference::partial(self.schema_name.clone(), name.to_string()),
-        )
-        .await
-        .map(|table| {
-            let table = table.with_dialect(Arc::clone(&self.dialect));
-            Some(Arc::new(table) as Arc<dyn TableProvider>)
-        })
-        .map_err(provider_error)
-    }
-
-    fn table_exist(&self, name: &str) -> bool {
-        self.tables.contains_key(name)
-    }
-
-    async fn table_type(&self, name: &str) -> DataFusionResult<Option<TableType>> {
-        Ok(self.tables.get(name).copied())
-    }
-}
 
 fn registered_source_for_catalog(
     common: &SourceManifestCommon,
@@ -515,14 +321,6 @@ fn database_relation_inventory(relations: &[DatabaseRelation]) -> Vec<Registered
             search_limits: None,
         })
         .collect()
-}
-
-fn provider_error(error: impl std::error::Error + Send + Sync + 'static) -> DataFusionError {
-    DataFusionError::External(Box::new(error))
-}
-
-fn boxed_provider_error(error: Box<dyn std::error::Error + Send + Sync>) -> DataFusionError {
-    DataFusionError::External(error)
 }
 
 #[cfg(test)]
